updated:renamed the js files
